requirements.txt
annotated-types==0.7.0
anyio==4.11.0
arrow==1.3.0
asgiref==3.10.0
audioop-lts==0.2.2
audioread==3.0.1
binaryornot==0.4.4
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cookiecutter==2.6.0
decorator==5.2.1
Django==5.2.7
django-tailwind==4.2.0
djangorestframework==3.16.1
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-genai==1.46.0
google-generativeai==0.8.5
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
honcho==2.0.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgpack==1.1.2
numba==0.62.1
numpy==2.3.4
packaging==25.0
platformdirs==4.5.0
pooch==1.8.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
sqlparse==0.5.3
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
tenacity==9.1.2
text-unidecode==1.3
threadpoolctl==3.6.0
tqdm==4.67.1
types-python-dateutil==2.9.0.20251008
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
waitress==3.0.2
websockets==15.0.1
